Move constants.ts to lib folder
